feat(core): stub out core API, define shared types
The core crate is intended to provide types that will be useful in both
processes: the http server and the worker.

This is also a convenient place to try and encapsulate the storage
concerns (i.e, database access). Ideally we'd be able to hide much of
the database dependency types/functions but in practice this is often
tricky or awkward. TBD if this can be achieved here.

Currently, I'm yet to actually pick a client library so the actual SQL
details are left unimplemented. The stubs should however be enough to
build out the downstream systems, with an allowance for some small
refactors as needed to tie it all together.