refactor: rename core->common, add initial db schema
Renaming the `core` crate to `common` became necessary due to some
integration pain in the `server` crate. Effectively, the prescribed way
to handle database connections bound to a given request's lifecycle is
to use `State` and a custom `Extractor` to acquire a connection from a
pooler.

The issue I am working around is the way this extractor is written, per
axum, dictates the use of `async-trait` which (apparently?!) will break
when there's a top-level item named `core`. Likely my crate was
shadowing <https://doc.rust-lang.org/core/> which is a terrible mistake
on my part! Renaming solves this issue.

This diff also refactors the storage code around the now known SQL
client dependency (I picked `tokio-postgres`) and adds a function for
building my initial thinking for the db schema.