feat(common): implement many SQL queries
This diff updates CRUD helpers to actually run SQL, as well as adding
supporting code for converting various enums to/from string so we can
move them through SQL.

The notable omission from the SQL helpers is the function which will do
the row-locking. I'm leaving this one until the very end when I flesh
out the worker binary.

N.b. I think ideally we'd either be using a numeric coding for these
enums (which is less readable when looking at raw tables) and/or
leveraging `ToSql` and `FromSql` trait implementations to handle
conversion more automatically.

Manually producing string versions and parsing from strings for now, in
the interest of time.