feat(sever): implement task list filtering
Ran into a bit of trouble on this one. Essentially it seems hard to
build up dynamic (varied) bind parameter lists for a query given the API
exposed by tokio-postgres.

I settled for (and actually this impl looks fine, for now) matching on
the pair of filters currently available and wrote a distinct query for
each combination (both filters, either one, none). As new filters are
added this will fast hit a point where this is unacceptable - another
approach will be needed to reduce maintenance burden.